We have a tool to run a body only once, use it
`dosync` does not create an exclusion context on executing its body,
so it's possible for two threads to enter the `dosync` and try to run
`eval-fn` independently. Only one of them will win and get into
`ref-map`, but there's still two outbound requests. This is bad.

Delay actually promises exactly once execution and deferred execution,
which are the two things we want. Makes sure that locks are localized
on individual bodies and we don't e.g. lock on the entire map.